% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consensus.tree}
\alias{consensus.tree}
\title{Consensus Tree of Extant Parasitoids}
\format{Phylogenetic tree in the \code{phylo} format defined in the
  \code{ape} package.}
\usage{
consensus.tree
}
\description{
A majority-rule consensus tree of a Bayesian analysis of a morphological
  dataset including extant and fossil parasitoid wasps, with the fossil
  terminals dropped. See Klopfstein & Spasojevic 2018 for details.
}
\references{
Klopfstein & Spasojevic (2018)
}
\keyword{datasets}
