% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogue.plot.R
\name{expand.state.space}
\alias{expand.state.space}
\title{Expand state space}
\usage{
expand.state.space(alignment, nb.states, return.as.list = F,
  log.name = "StateChanges_expanded.log")
}
\arguments{
\item{alignment}{An alignment of characters of type "standard", in the
form of either a list of the same length as there are taxa in the
alignment, or as a matrix with as many rows as taxa and as many columns
as characters.}

\item{nb.states}{a single value or vector of type \code{numeric} giving the
number of states each characters should contain. These number have to be
equal or higher than the current number of states in each character;
otherwise the matrix is returned unchanged.}

\item{return.as.list}{a logical specifying whether the new alignment should
be returned as a list (if \code{TRUE}) or as a matrix (if \code{FALSE}).}

\item{log.name}{name of the log file to which details about the conducted
changes are written.}
}
\value{
\code{expand.state.space} returns either a list with each element
  representing a taxon (if \code{return.as.list} equals \code{TRUE}), or a
  matrix with each row representing a taxon.
}
\description{
Adding unobserved states to an alignment of standard characters.
}
\details{
The "standard" data type as it is defined in the NEXUS format is used to
  represent discrete characters such as they result from morphological
  examination. They are often represented by numbers as state labels ("0",
  "1", "2", etc.) and their evolution can be modelled using Lewis' (2001)
  likelihood approach.
Phylogenetic inference programs such as MrBayes that can handle such
  discrete characters, assume that the true number of states that are
  possible in each character is given by the largest number observed in
  the corresponding column of the data matrix. This function adds empty
  states to ascertain that the largest state label equals the target
  number of states as specified in \code{nb.states}. Changes to state labels
  are recorded in a log file.
The target number of states can for instance stem from theoretical
  considerations or from character states in a more inclusive matrix.
  \code{nb.states}) can either contain a single value, which is then assumed
  to be the same for all the characters, or a vector of the same length as
  the number of characters in the matrix. Only single-digit numbers ("0"-"9")
  are currently allowed, a restriction also imposed by software such as
  MrBayes.
}
\examples{
## compare the state labels
data(orig.morpho)
data(tot.nb.states)
expanded.morpho <- expand.state.space(orig.morpho, nb.states = tot.nb.states,
  return.as.list = TRUE)
str(unlist(lapply(orig.morpho, "[", 36)))
str(unlist(lapply(expanded.morpho, "[", 36)))
}
\references{
Lewis (2001) A likelihood approach to estimating phylogeny from discrete
  morphological character data. Systematic Biology, 50, 913-925.
}
