% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fossil.trees}
\alias{fossil.trees}
\title{Bayesian Trees of Fossil and Extant Parasitoids}
\format{Phylogenetic trees in the \code{multiphylo} format defined in the
  \code{ape} package.}
\usage{
fossil.trees
}
\description{
A set of 100 phylogenetic trees from a Bayesian analysis of a morphological
  dataset including extant and fossil parasitoid wasps. See Klopfstein &
  Spasojevic 2018 for details.
}
\references{
Klopfstein & Spasojevic (2018)
}
\keyword{datasets}
