% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogue.plot.R
\name{get.edge.formats}
\alias{get.edge.formats}
\title{Get Edge Formatting for Rogue Plots}
\usage{
get.edge.formats(tree, placements, zero.col = "black", col = paste("grey",
  seq(from = 95, to = 15, by = -8), sep = ""), zero.lwd = 1, lwd = 3,
  min.prob = 0.01)
}
\arguments{
\item{tree}{an object of class \code{phylo}.}

\item{placements}{a list containing, for each rogue taxon, the sister
groups of the rogue and their frequencies in the reference trees. For
details about the format see the help of the
\code{get.rogue.placement}) function.}

\item{zero.col}{colour used to reflect a placement probability below
\code{min.prob}.}

\item{col}{a vector with (at least) 10 colours which reflect placement
probabilities between \code{min.prob} and 1.}

\item{zero.lwd}{line width used to reflect a placement probability below
\code{min.prob}. Defaults to 1.}

\item{lwd}{line width used to reflect a placement probability above
\code{min.prob}. Defaults to 3.}

\item{min.prob}{minimum probability of attachment, below which an edge is
considered as zero for plotting. Defaults to 1\% of reference trees.}
}
\value{
Returns a list containing, for each rogue taxon, the horizontal and vertical
  colour vectors and horizontal and vertical line widths. Beware that
  the lengths of the horizontal vectors corresponds to the number of edges
  in the tree, while the vertical ones contain as many elements as the tree
  has internal nodes.
}
\description{
Given a tree and a list of taxon placements (as generated by the
  \code{\link{get.rogue.placement}}) function), this function generates a
  list with colours and line widths for the horizontal and vertical edges of
  a phylogram.
}
\details{
\code{get.edge.formats} compares a list containing rogue taxon
  placement information (in the format generated by
  \code{\link{get.rogue.placement}}) with the tree to plot and generates a
  list of formating vectors for the horizontal and vertical branches of a
  phylogram. Both colour and line type of an edge can be set according to
  the rogue taxon placement probabilites.
}
\examples{
## example using random trees
reftrees <- ape::rmtree(5, 21, tip.label = c("rogue.taxon", paste("t",
  1:20, sep = "")))
reftrees <- c(rep(list(reftrees[[1]]), 60), rep(list(reftrees[[2]]),
  25), rep(list(reftrees[[3]]), 10), reftrees)
tree <- reftrees[[1]]
placements <- get.rogue.placement(tree, reftrees, rogues =
  "rogue.taxon", outgroup = "t1")
edge.formats <- get.edge.formats(tree, placements, zero.col = "black",
  col = paste("grey", seq(from = 95, to = 20, by = -8), sep = ""),
  zero.lwd = 1, lwd = 3, min.prob = 0.01)
str(edge.formats)

## example from Klopfstein & Spasojevic 2018
data(consensus.tree)
data(fossil.trees)
data(taxon.lists)
placements <- get.rogue.placement(consensus.tree, fossil.trees,
  rogues = taxon.lists$fossils[7], outgroup = taxon.lists$outgroup[1])
edge.formats <- get.edge.formats(tree, placements, zero.col = "black",
  col = paste("grey", seq(from = 95, to = 20, by = -8), sep = ""),
  zero.lwd = 1, lwd = 3, min.prob = 0.01)
str(edge.formats)
}
\references{
Klopfstein, S., Spasojevic, T.(2018) ...
}
