% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogue.plot.R
\name{get.rogue.placement}
\alias{get.rogue.placement}
\title{Get Rogue Taxon Placements}
\usage{
get.rogue.placement(tree, reftrees, rogues, outgroup = 1)
}
\arguments{
\item{tree}{an object of class \code{phylo}.}

\item{reftrees}{an object of class \code{multiphylo} containing a set of
treesfrom a bootstrap or Bayesian analysis; they all have to contain the
same tips, and the same tips as \code{tree} PLUS any rogue taxa to analyze.}

\item{rogues}{a vector of mode character containing the names of the tip or
tips whose placement is to be calculated.}

\item{outgroup}{a vector of mode numeric or character specifying the new
outgroup.}
}
\value{
Returns a list whose elements correspond to a dataframe for each rogue,
  which contain comma-separated lists of sister taxa, their frequency among
  the reference trees, whether or not the sister group is monophyletic in
  the consensus tree, and finally its node index in the tree to plot. If a
  sister group does not appear in the tree, then the node corresponding to
  the most recent common ancestor is recorded.
}
\description{
Function calculating the frequency of different placements of one or several
rogue taxa in a set of reference trees.
}
\details{
This function lists all groups appearing as sister groups to the rogue taxon
  (or taxa) in any of the reference trees, their respective frequencies,
  and whether or not they are present in the tree that is to be plotted.

Make sure that the reference trees contain exactly the taxa in the tree,
  plus the rogue taxon or taxa.

The reference trees are rooted before analysis, using either the specified
  outgroup or the first taxon in the tiplabels. A vector of outgroup taxa,
  either in numeric or character format, can be passed, in which case the
  function checks whether the specified outgroup is monophyletic. If not,
  a warning is issued and the first taxon among the outgroups is used to
  root the tree.
}
\examples{
## example using random trees
reftrees <- ape::rmtree(5, 21, tip.label = c("rogue.taxon", paste("t",
  1:20, sep = "")))
reftrees <- c(rep(list(reftrees[[1]]), 60), rep(list(reftrees[[2]]), 25),
  rep(list(reftrees[[3]]), 10), reftrees)
tree <- ape::root(reftrees[[1]], outgroup = "t1")
placements <- get.rogue.placement(tree, reftrees, rogues = "rogue.taxon",
  outgroup = "t1")
str(placements)

## example from Klopfstein & Spasojevic 2018
data(consensus.tree)
data(fossil.trees)
data(taxon.lists)
placements <- get.rogue.placement(consensus.tree, fossil.trees,
  rogues = c(taxon.lists$fossils[1], taxon.lists$fossils[7]),
  outgroup = taxon.lists$outgroup[1])
str(placements)

}
\references{
Klopfstein, S., Spasojevic, T. (2018). Illustrating phylogenetic placement
of fossils using RoguePlots: An example from ichneumonid parasitoid wasps
(Hymenoptera, Ichneumonidae) and an extensive morphological matrix. bioRxiv
https://doi.org/10.1101/425090
}
