% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orig.morpho}
\alias{orig.morpho}
\title{The Original Dataset from Klopfstein & Spasejovic (2018)}
\format{List with as elements for each taxon a vector of mode
  \code{character} containing the coding of 222 morphological characters.
  The names of the list items correspond to the taxa in the analysis.}
\usage{
orig.morpho
}
\description{
Morphological matrix in its original form as it was used in the parasitoid
  fossil study. See Klopfstein & Spasojevic 2018 for details.
}
\references{
Klopfstein & Spasojevic (2018)
}
\keyword{datasets}
