% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogue.plot.R
\name{prob.scale}
\alias{prob.scale}
\title{Create Scale for Probabilities}
\usage{
prob.scale(col, breaks, title = "probability", cex = par("cex"))
}
\arguments{
\item{col}{a vector of mode character giving the colours used to reflect
probabilities. At least two colours have to be passed.}

\item{breaks}{(optional): the limits of the categories of the scale. If
\code{breaks} are not provided, they are spaced equally from zero to one,
in order to match the number of colours passed in \code{col}.}

\item{title}{a title for the scale. Pass \code{'""'} to suppress the title.}

\item{cex}{a numeric value giving the scaling factor of the axis labels
(Character EXpansion). The default is to take the current value from the
graphical parameters. If printed, the title of the scale is chosen to be
1.2x as large.}
}
\description{
This function plots a scale from zero to one next to a vertical colour code.
}
\details{
\code{prob.scale} prints a probability scale using the colour vector
  passed to it. The scale is not added to the current plot but printed
  separately. Use \code{\link{layout}} or similar to create subplot areas if
  intending to plot the scale next to a graph.
}
\examples{
## example of an evenly spaced grey scale:
prob.scale(col = paste("grey", 10*(9:2), sep = ""), cex = 0.9)

## example of an evenly spaced rainbow scale:
prob.scale(col = rainbow(n = 8), cex = 0.9)

## example of an unevenly scale with user-defined breaks:
prob.scale(col = c("grey80", rainbow(n = 5, start = 0.35, end = 0.7)),
  breaks = c(0, 0.05, seq(from = 0.2, to = 1.0, by = 0.2)), cex = 0.9)
}
