% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogue.plot.R
\name{read.nexus.morpho}
\alias{read.nexus.morpho}
\title{Read NEXUS-formated morphological data with polymorphisms}
\usage{
read.nexus.morpho(file, return.as.list = TRUE)
}
\arguments{
\item{file}{a file name specified either by a variable of mode character or
by a double-quoted string.}

\item{return.as.list}{a logical specifying whether the alignment should be
returned as a list (if \code{TRUE}) or as a matrix (if \code{FALSE}).}
}
\value{
\code{read.nexus.morpho} returns either a list with each entry representing
  a taxon (if \code{return.as.list} equals \code{TRUE}), or a matrix with
  each row representing a taxon.
}
\description{
This function allows reading in 'standard' data in NEXUS format that can
  include polymorphisms.
}
\details{
Morphological data in NEXUS format is read in from the first data block in
  the specified file (all following data blocks are ignored). The data has
  to be on a single line, after the taxon identifier followed by spaces or
  tabs - the 'interleaved' data format is not supported.
Polymorphic character states have to be specified in brackets and separated
  by commas, for example \code{(0,1,4)}.
}
\examples{
## first writing the dataset from Klopfstein & Spasojevic (2018) to a file
## using the \\code{write.nexus.data} function from the package \\code{ape},
## then reading it in using the current function which handles polymorphies.
data(orig.morpho)
ape::write.nexus.data(orig.morpho, file = "morpho.temp", format = "standard",
   interleaved = FALSE)
morpho <- read.nexus.morpho(file = "morpho.temp", return.as.list = TRUE)
str(morpho)
unlink("morpho.temp")
}
\references{
Klopfstein, S., Spasojevic, T. (2018).
}
