% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogue.plot.R
\name{restrict.state.space}
\alias{restrict.state.space}
\title{Restrict state space}
\usage{
restrict.state.space(alignment, return.as.list = F,
  log.name = "StateChanges_restrict.log")
}
\arguments{
\item{alignment}{An alignment of characters of type "standard", in the
form of either a list of the same length as there are taxa in the
alignment, or as a matrix with as many rows as taxa and as many columns
as characters.}

\item{return.as.list}{a logical specifying whether the new alignment should
be returned as a list (if \code{TRUE}) or as a matrix (if \code{FALSE}).}

\item{log.name}{Name of the log file to which a summary of the changes
is written. Defaults to"StateChanges_restrict.log".}
}
\value{
\code{restrict.state.space} returns either a list with each element
  representing a taxon (if \code{return.as.list} equals \code{TRUE}), or a
  matrix with each row representing a taxon.
}
\description{
Removing unobserved states from an alignment of standard characters,
  preserving polymorphisms.
}
\details{
The "standard" data type as it is defined in the NEXUS format is used to
  represent discrete characters such as they result from morphological
  examination. They are often represented by numbers as state labels ("0",
  "1", "2", etc.) and their evolution can be modelled using Lewis' (2001)
  likelihood approach.
Phylogenetic inference programs such as MrBayes that can handle such
  discrete characters, assume that the true number of states that are
  possible in each character is given by the largest number observed in
  the corresponding column of the data matrix. This function removes empty
  states and recodes the remaining ones to ascertain that the smallest state
  label is "0" and the largest equals the number of states minus 1. Changes
  to state labels are recorded in a log file.
}
\examples{
data(orig.morpho)
restr.morpho <- restrict.state.space(orig.morpho, return.as.list = TRUE)
str(unlist(lapply(orig.morpho, "[", 222)))
str(unlist(lapply(restr.morpho, "[", 222)))
}
\references{
Lewis (2001) A likelihood approach to estimating phylogeny from discrete
  morphological character data. Systematic Biology, 50, 913-925.
}
