% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxon.lists}
\alias{taxon.lists}
\title{A List of Taxon Groups in the Parasitoids Dataset}
\format{List containing two vectors of mode \code{character}.
\describe{
  \item{outgroup}{outgroup taxa to root the trees in the parasitoid dataset}
  \item{fossils}{fossil taxa in the dataset}
  }}
\usage{
taxon.lists
}
\description{
List of two vectors each containing important taxon sets in the parasitoid
  data. See Klopfstein & Spasojevic 2018 for details.
}
\references{
Klopfstein & Spasojevic (2018)
}
\keyword{datasets}
