% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tot.nb.states}
\alias{tot.nb.states}
\title{Expanded Number of States}
\format{vector with as many integers as morphological characters.}
\usage{
tot.nb.states
}
\description{
Vector containing for each character in the morphological matrix the number
  of states observed in a matrix with more taxa. Used to inform the total
  number of states in \code{expand.state.space}.
}
\references{
Klopfstein, S., Spasojevic, T. (2018). Illustrating phylogenetic placement
of fossils using RoguePlots: An example from ichneumonid parasitoid wasps
(Hymenoptera, Ichneumonidae) and an extensive morphological matrix. bioRxiv
https://doi.org/10.1101/425090
}
\keyword{datasets}
